requirements.txt
django~=3.1.7
gunicorn
django-heroku
markdown
django-filter
djangorestframework~=3.12.2
requests~=2.25.1